+++ CHANGES.rst
@@ -7,6 +7,7 @@ Improvements:
  * NSE: Make extension reusable (#3326).
  * Strings: Use you instead of display name on notice events (#3282).
  * Third-party licences: Add license for FlowCommoniOS (#3415).
+ * Lazy-loading: Remove lazy loading labs setting, enable it by default (#3389).
  
 Bug fix:
  * Xcode11: Fix content change error when dragging start chat page (PR #3075).